moved append_video_resize_filter to ffmpeg class for intuitive reasons
